Move DurationStatistics to core module
