+++ java/core/src/test/java/sleeper/core/util/DurationStatisticsTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.status.report.job;
+package sleeper.core.util;
 
 import org.junit.jupiter.api.Test;
